Semi working Database implementation. to be redone soon

+++ migrations/2023-09-16-201101_core_users/up.sql
@@ -1,9 +1,9 @@
 -- Your SQL goes here
 CREATE TABLE IF NOT EXISTS "core_users" (
-	"user_id"			INTEGER,
+	"user_id"			BIGINT,
 	"user_name"			TEXT NOT NULL,
 	"user_nickname"		TEXT NOT NULL,
-	"user_created_at"	INTEGER NOT NULL,
-	"user_role"			INTEGER NOT NULL,
+	"user_created_at"	BIGINT NOT NULL,
+	"user_role"			BIGINT NOT NULL,
 	PRIMARY KEY("user_id")
 );
