+++ migrations/2023-09-16-201241_moderation_messages/up.sql
@@ -1,6 +1,6 @@
 -- Your SQL goes here
 CREATE TABLE IF NOT EXISTS "moderation_messages" (
-	"message_id"	INTEGER,
+	"message_id"	BIGINT,
 	"message_link"	TEXT NOT NULL,
 	PRIMARY KEY("message_id")
 );
