+++ migrations/2023-09-16-201250_moderation_notes/up.sql
@@ -1,10 +1,10 @@
 -- Your SQL goes here
 CREATE TABLE IF NOT EXISTS "moderation_notes" (
-	"note_id"	INTEGER,
-	"target_id"	INTEGER NOT NULL,
-	"moderator_id"	INTEGER NOT NULL,
+	"note_id"	BIGINT,
+	"target_id"	BIGINT NOT NULL,
+	"moderator_id"	BIGINT NOT NULL,
 	"note"	TEXT NOT NULL,
-	"message_reference"	INTEGER,
+	"message_reference"	BIGINT,
 	FOREIGN KEY("moderator_id") REFERENCES "core_users"("user_id"),
 	FOREIGN KEY("target_id") REFERENCES "core_users"("user_id"),
 	FOREIGN KEY("message_reference") REFERENCES "moderation_message_references"("message_reference"),
