+++ migrations/2023-09-16-201257_moderation_bans/up.sql
@@ -1,10 +1,10 @@
 -- Your SQL goes here
 CREATE TABLE IF NOT EXISTS "moderation_bans" (
-	"ban_id"	INTEGER,
-	"target_id"	INTEGER NOT NULL,
-	"moderator_id"	INTEGER NOT NULL,
+	"ban_id"	BIGINT,
+	"target_id"	BIGINT NOT NULL,
+	"moderator_id"	BIGINT NOT NULL,
 	"ban_reason"	TEXT,
-	"ban_duration"	NUMERIC,
+	"ban_duration"	BIGINT,
 	FOREIGN KEY("target_id") REFERENCES "core_users"("user_id"),
 	FOREIGN KEY("moderator_id") REFERENCES "core_users"("user_id"),
 	PRIMARY KEY("ban_id")
