+++ migrations/2023-09-16-203354_moderation_message_references/up.sql
@@ -1,8 +1,8 @@
 -- Your SQL goes here
 CREATE TABLE IF NOT EXISTS "moderation_message_references" (
-	"message_reference"	INTEGER,
-	"entry_id"	INTEGER,
-	"message_id"	INTEGER NOT NULL,
+	"message_reference"	BIGINT,
+	"entry_id"	BIGINT,
+	"message_id"	BIGINT NOT NULL,
 	"note"	TEXT,
 	FOREIGN KEY("message_id") REFERENCES "moderation_messages"("message_id"),
 	PRIMARY KEY("message_reference","entry_id")
